extract shared logic
update test plan selectors logic
