+++ packages/allure-vitest/src/reporter.ts
@@ -16,6 +16,7 @@ import {
   getSuitesLabels,
 } from "allure-js-commons";
 import { ALLURE_SKIPPED_BY_TEST_PLAN_LABEL } from "allure-js-commons/internal";
+import { getTestFullName, getSuitePath } from "./utils.js";
 
 export interface AllureReporterOptions {
   testMode?: boolean;
@@ -55,23 +56,6 @@ export default class AllureReporter implements Reporter {
         url,
       };
     });
-  }
-
-  private getSuitePath(task: Task): string[] {
-    const path = [];
-    let currentSuite: Suite | undefined = task.suite;
-
-    while (currentSuite) {
-      // root suite has no name and shouldn't be included to the path
-      if (!currentSuite.name) {
-        break;
-      }
-
-      path.unshift(currentSuite.name);
-      currentSuite = currentSuite.suite;
-    }
-
-    return path;
   }
 
   onInit(vitest: Vitest) {
@@ -132,13 +116,13 @@ export default class AllureReporter implements Reporter {
     const links = currentTest.links ? this.processMetadataLinks(currentTest.links) : [];
     const labels: Label[] = [].concat(currentTest.labels || []).concat(titleMetadata.labels);
     const test = parent.startTest(testDisplayName);
-    const suitePath = this.getSuitePath(task);
+    const suitePath = getSuitePath(task);
     const normalizedTestPath = normalize(task.file.filepath.replace(this.rootDir, ""))
       .replace(/^\//, "")
       .split("/")
       .filter((item: string) => item !== basename(task.file.filepath));
 
-    test.fullName = `${task.file.name}#${suitePath.concat(task.name).join(" ")}`;
+    test.fullName = getTestFullName(task);
     test.applyMetadata({
       ...currentTest,
       labels,
@@ -149,7 +133,7 @@ export default class AllureReporter implements Reporter {
     test.addLabel(LabelName.THREAD, ALLURE_THREAD_NAME || pid.toString());
     test.addLabel(LabelName.HOST, ALLURE_HOST_NAME || hostname.toString());
 
-    getSuitesLabels(this.getSuitePath(task)).forEach((label) => {
+    getSuitesLabels(suitePath).forEach((label) => {
       test.addLabel(label.name, label.value);
     });
