+++ packages/allure-vitest/src/setup.ts
@@ -1,6 +1,7 @@
 import { type TaskContext, afterAll, afterEach, beforeAll, beforeEach } from "vitest";
 import { LabelName, TestPlanV1, extractMetadataFromString, parseTestPlan } from "allure-js-commons";
 import { ALLURE_SKIPPED_BY_TEST_PLAN_LABEL } from "allure-js-commons/internal";
+import { getTestFullName } from "./utils.js";
 import { bindAllureApi } from "./index.js";
 
 const existsInTestPlan = (ctx: TaskContext, testPlan?: TestPlanV1) => {
@@ -10,15 +11,14 @@ const existsInTestPlan = (ctx: TaskContext, testPlan?: TestPlanV1) => {
 
   const {
     name: testName,
-    file: { name: testFileName },
   } = ctx.task;
+  const testFullName = getTestFullName(ctx.task);
   const { labels } = extractMetadataFromString(testName);
   const allureIdLabel = labels.find(({ name }) => name === LabelName.ALLURE_ID);
 
   return testPlan.tests.some(({ id, selector = "" }) => {
     const idMatched = id ? String(id) === allureIdLabel?.value : false;
-    const splittedSelector = selector.split("#");
-    const selectorMatched = splittedSelector[0] === testFileName && splittedSelector[1] === testName;
+    const selectorMatched = selector === testFullName;
 
     return idMatched || selectorMatched;
   });
